requirements.txt
pdf2image
PyPDF2
tqdm
Pillow
pytesseract
numpy
opencv-python
